refactor: components 폴더 구조 변경 - 페이지별 폴더 및 common 폴더 추가
