+++ app/(root)/layout.jsx
@@ -1,6 +1,5 @@
-import Footer from '@/components/Footer';
-import Header from '@/components/Header';
-import React from 'react';
+import Footer from '@/components/common/Footer';
+import Header from '@/components/common/Header';
 
 function Rootlayout({ children }) {
   return (
